move window code into a separate module and format code
